Avoid internal signing by lopping off chain top
Removing the top of the chain makes it possible to avoid checking
the trust-anchor signatures, and thus none of the temporary CAs
above it need to be signed or to have signed the TA.

Add verify callback to the test program, and add support for full
cert/key data (mtype 0) via empty digest name.

More detailed usage message.

Disable SSLv2 when enabling SNI.